Add floating button to navigation bar
For some reason, when I tried to a fragment to activity_main.xml,
the application crashed. It gave me the following error: 'Error inflating class fragment'.

Since the floagin button is in our mockups, I decided to still add it to git but not in
the master branch so our teacher can still evaluate what we have done.